requirements.txt
flask
gunicorn
sentry-sdk[flask]
psycopg2-binary
requests
curl_cffi
python-telegram-bot
